Adding initial "mover" helper class. connect segfaults.
Connect segfaults when combining more than one function, as the
intermediate Sink is passed to the next call as an rvalue reference.
The hope is that this sink can then be moved into another thread using
move semantics. Currently since it is used by reference, the code
segfaults as it is immediately destroyed after the connect call
completes.

Unfortunately lambdas don't capture by move, so I am trying to come up
with a hack on how to do it. Once that works, segfaults should be gone.